@@ -2,13 +2,16 @@
     "name": "silktide/semrush-api",
     "description": "SEMrush API client",
     "require": {
-        "php": ">=5.6.0",
+        "php": ">=7.0",
         "guzzlehttp/guzzle": "^6.1",
-        "psr/log": "^1.0"
+        "psr/log": "^1.0",
+        "psr/simple-cache": "^1.0",
+        "cache/array-adapter": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.*",
-        "codeclimate/php-test-reporter": "0.1.*"
+        "silktide/pharmacist": "^1.0",
+        "phpunit/phpunit": "^6.5",
+        "monolog/monolog": "^1.23"
     },
     "license": "GPL-2.0+",
     "authors": [
