@@ -4,9 +4,9 @@
 namespace Silktide\SemRushApi\Test\Data;
 
 use Silktide\SemRushApi\Data\Type;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class TypeTest extends PHPUnit_Framework_TestCase {
+class TypeTest extends TestCase {
 
     /**
      * Test getting the types
