@@ -3,10 +3,10 @@
 namespace Silktide\SemRushApi\Test\Model\Factory;
 
 use Silktide\SemRushApi\Model\Factory\RequestFactory;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Silktide\SemRushApi\Data\Type;
 
-class RequestFactoryTest extends PHPUnit_Framework_TestCase {
+class RequestFactoryTest extends TestCase {
 
     /**
      * @var RequestFactory
@@ -24,6 +24,7 @@ class RequestFactoryTest extends PHPUnit_Framework_TestCase {
     public function testCreate()
     {
         $this->instance->create(Type::TYPE_DOMAIN_RANKS, ['key' => 'a key', 'domain' => 'domain.com']);
+        $this->assertTrue(true);
     }
 
 
