@@ -6,10 +6,10 @@ namespace Silktide\SemRushApi\Test\Model\Factory;
 use Silktide\SemRushApi\Model\Result;
 use Silktide\SemRushApi\Model\Row;
 use Silktide\SemRushApi\Test\ResponseExample\ResponseExampleHelper;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 use Silktide\SemRushApi\Model\Factory\ResultFactory;
 
-class ResultFactoryTest extends PHPUnit_Framework_TestCase {
+class ResultFactoryTest extends TestCase {
 
     /**
      * @var ResultFactory
@@ -23,7 +23,7 @@ class ResultFactoryTest extends PHPUnit_Framework_TestCase {
      */
     public function setup()
     {
-        $rowFactory = $this->getMock('Silktide\SemRushApi\Model\Factory\RowFactory');
+        $rowFactory = $this->createMock('Silktide\SemRushApi\Model\Factory\RowFactory');
         $row = $this->getMockBuilder('Silktide\SemRushApi\Model\Row')->disableOriginalConstructor()->getMock();
         $rowFactory->expects($this->any())->method('create')->willReturn($row);
         $this->instance = new ResultFactory($rowFactory);
