@@ -6,10 +6,11 @@ namespace Silktide\SemRushApi\Test\Model;
 use Silktide\SemRushApi\Data\Column;
 use Silktide\SemRushApi\Data\Database;
 use Silktide\SemRushApi\Data\Type;
+use Silktide\SemRushApi\Model\Exception\InvalidOptionException;
 use Silktide\SemRushApi\Model\Request;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase;
 
-class RequestTest extends PHPUnit_Framework_TestCase {
+class RequestTest extends TestCase {
 
     protected $key = 'testkey';
     protected $domain = 'domain.com';
@@ -73,61 +74,71 @@ class RequestTest extends PHPUnit_Framework_TestCase {
 
     public function testRequestWithMissingOption()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', 'Missing option(s) [key, domain] which are required for request [domain_ranks]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('Missing option(s) [key, domain] which are required for request [domain_ranks]');
         new Request(Type::TYPE_DOMAIN_RANKS, []);
     }
 
     public function testRequestWithInvalidColumnData()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', 'Invalid [export_columns](s) [Blah] passed for request [domain_ranks]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('Invalid [export_columns](s) [Blah] passed for request [domain_ranks]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => $this->key, 'domain' => $this->domain, 'export_columns' => ['Blah', 'At']]);
     }
 
     public function testRequestWithInvalidColumnValue()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[export_columns] is expected to be array for [domain_ranks]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[export_columns] is expected to be array for [domain_ranks]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => $this->key, 'domain' => $this->domain, 'export_columns' => 'blah']);
     }
 
     public function testRequestWithInvalidOption()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', 'Invalid option(s) [something] passed for request [domain_ranks]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('Invalid option(s) [something] passed for request [domain_ranks]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['something' => 'bad']);
     }
 
     public function testRequestWithInvalidDate()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[display_date] was not a valid date [20301215]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[display_date] was not a valid date [20301215]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => $this->key, 'domain' => $this->domain, 'display_date' => "20301215"]);
     }
 
     public function testRequestWithInvalidBoolean()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[export_escape] was not 1 or 0 [wrong]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[export_escape] was not 1 or 0 [wrong]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => $this->key, 'domain' => $this->domain, 'export_escape' => 'wrong']);
     }
 
     public function testRequestWithInvalidDomain()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[domain] was not a valid domain [not a domain]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[domain] was not a valid domain [not a domain]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => $this->key, 'domain' => "not a domain"]);
     }
 
     public function testRequestWithInvalidInteger()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[display_limit] was not an integer [beep]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[display_limit] was not an integer [beep]');
         new Request(Type::TYPE_DOMAIN_RANK_HISTORY, ['key' => $this->key, 'domain' => $this->domain, 'display_limit' => "beep", 'database' => 'uk']);
     }
 
     public function testRequestWithInvalidDatabase()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[database] was not a database [rubbish]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[database] was not a database [rubbish]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => $this->key, 'domain' => $this->domain, 'database' => 'rubbish']);
     }
 
     public function testRequestWithInvalidKey()
     {
-        $this->setExpectedException('Silktide\SemRushApi\Model\Exception\InvalidOptionException', '[key] was not a string [12]');
+        $this->expectException(InvalidOptionException::class);
+        $this->expectExceptionMessage('[key] was not a string [12]');
         new Request(Type::TYPE_DOMAIN_RANKS, ['key' => 12, 'domain' => $this->domain]);
     }
 
