@@ -2,9 +2,9 @@
 
 namespace Silktide\SemRushApi;
 
-use Guzzle\Http\Exception\BadResponseException;
+use GuzzleHttp\Exception\BadResponseException;
 use GuzzleHttp\RequestOptions;
-use Silktide\SemRushApi\Cache\CacheInterface;
+use Psr\SimpleCache\CacheInterface;
 use Silktide\SemRushApi\Data\Type;
 use Silktide\SemRushApi\Helper\ResponseParser;
 use Silktide\SemRushApi\Helper\UrlBuilder;
@@ -31,14 +31,14 @@ class Client
     protected $requestFactory;
 
     /**
-     * @var ResultFactory
-     */
-    protected $resultFactory;
-
-    /**
      * @var ResponseParser
      */
     protected $responseParser;
+
+    /**
+     * @var ResultFactory
+     */
+    protected $resultFactory;
 
     /**
      * @var UrlBuilder
@@ -51,11 +51,6 @@ class Client
     protected $guzzle;
 
     /**
-     * @var CacheInterface
-     */
-    protected $cache;
-
-    /**
      * @var int
      */
     protected $connectTimeout = 15;
@@ -65,6 +60,20 @@ class Client
      */
     protected $timeout = 15;
 
+    /**
+     * @var CacheInterface|null
+     */
+    protected $cache;
+
+    /**
+     * @var \DateInterval|null
+     */
+    protected $cacheLifetime;
+
+    /**
+     * @var string
+     */
+    protected $cachePrefix = "";
 
     /**
      * Construct a client with API key
@@ -76,14 +85,29 @@ class Client
      * @param UrlBuilder $urlBuilder
      * @param GuzzleClient $guzzle
      */
-    public function __construct($apiKey, RequestFactory $requestFactory, ResultFactory $resultFactory, ResponseParser $responseParser, UrlBuilder $urlBuilder, GuzzleClient $guzzle)
+    public function __construct($apiKey, RequestFactory $requestFactory, ResponseParser $responseParser, ResultFactory $resultFactory, UrlBuilder $urlBuilder, GuzzleClient $guzzle)
     {
         $this->apiKey = $apiKey;
         $this->requestFactory = $requestFactory;
-        $this->resultFactory = $resultFactory;
         $this->responseParser = $responseParser;
+        $this->resultFactory = $resultFactory;
         $this->urlBuilder = $urlBuilder;
         $this->guzzle = $guzzle;
+    }
+
+    /**
+     * @param CacheInterface $cache
+     * @param \DateInterval $cacheLifetime
+     * @param string $cachePrefix
+     */
+    public function setCache(CacheInterface $cache, \DateInterval $cacheLifetime = null, string $cachePrefix = "")
+    {
+        $this->cache = $cache;
+        if ($cacheLifetime === null) {
+            $cacheLifetime = new \DateInterval("P30D");
+        }
+        $this->cacheLifetime = $cacheLifetime;
+        $this->cachePrefix = $cachePrefix;
     }
 
     /**
@@ -116,15 +140,6 @@ class Client
     public function setTimeout($timeout)
     {
         $this->timeout = $timeout;
-    }
-
-
-    /**
-     * @param CacheInterface $cache
-     */
-    public function setCache(CacheInterface $cache)
-    {
-        $this->cache = $cache;
     }
 
     /**
@@ -238,43 +253,21 @@ class Client
     }
 
     /**
-     * Make the request
-     *
-     * @param string $type
-     * @param array $options
+     * @param $type
+     * @param $options
      * @return ApiResult
+     * @throws Exception
      */
     protected function makeRequest($type, $options)
     {
         try {
             $request = $this->requestFactory->create($type, ['key' => $this->apiKey] + $options);
-
-            // Attempt load from cache
-            if (isset($this->cache)) {
-                $result = $this->cache->fetch($request);
-            }
-
-            // Make request if not in cache
-            if (!isset($result)) {
-                $rawResponse = $this->makeHttpRequest($request);
-                $formattedResponse = $this->responseParser->parseResult($request, $rawResponse);
-                $result = $this->resultFactory->create($formattedResponse);
-            }
-
-            // Save to cache
-            if (isset($this->cache)) {
-                $this->cache->cache($request, $result);
-            }
-
-            return $result;
-
+            $rawResponse = $this->makeHttpRequest($request);
+            $formattedResponse = $this->responseParser->parseResult($request, $rawResponse);
+            return $this->resultFactory->create($formattedResponse);
         } catch (BadResponseException $e) {
-            $this->logBadResponse($e);
-        } catch (Exception $e) {
-            $this->logException($e);
+            throw $this->parseBadResponse($e);
         }
-
-        return $this->resultFactory->create([]);
     }
 
     /**
@@ -286,11 +279,45 @@ class Client
     protected function makeHttpRequest($request)
     {
         $url = $this->urlBuilder->build($request);
+
+        $cacheKey = $this->cachePrefix . md5($url);
+        if ($this->cache) {
+            if (!is_null($value = $this->cache->get($cacheKey))) {
+                return $value;
+            }
+        }
+
         $guzzleResponse = $this->guzzle->request('GET', $url, [
             RequestOptions::CONNECT_TIMEOUT => $this->connectTimeout,
             RequestOptions::TIMEOUT => $this->timeout
         ]);
-        return $guzzleResponse->getBody();
+
+        $value = $guzzleResponse->getBody();
+
+        if ($this->cache) {
+            $this->cache->set($cacheKey, $value, $this->cacheLifetime);
+        }
+
+        return $value;
+    }
+
+    /**
+     * @param BadResponseException $e
+     * @return \Exception
+     */
+    protected function parseBadResponse(BadResponseException $e)
+    {
+        $response = $e->getResponse();
+        $message = (string)$response->getBody();
+
+        if (!is_null($this->logger)) {
+            $this->logger->error("[SemRush API] " . $message, [
+                "StatusCode" => $response->getStatusCode(),
+                "URL" => (string)$e->getRequest()->getUri()
+            ]);
+        }
+
+        return new \Exception("[SemRush API] " . $message);
     }
 
     /**
