@@ -2,14 +2,12 @@
 
 namespace Silktide\SemRushApi;
 
-use GuzzleHttp\Client as Guzzle;
-use Silktide\SemRushApi\Cache\MemoryCache;
+use GuzzleHttp\Client as GuzzleClient;
 use Silktide\SemRushApi\Helper\ResponseParser;
 use Silktide\SemRushApi\Helper\UrlBuilder;
 use Silktide\SemRushApi\Model\Factory\RequestFactory;
 use Silktide\SemRushApi\Model\Factory\ResultFactory;
 use Silktide\SemRushApi\Model\Factory\RowFactory;
-use Silktide\SemRushApi\Cache\CacheInterface;
 
 abstract class ClientFactory
 {
@@ -19,25 +17,16 @@ abstract class ClientFactory
      * automatically created (for use by people without DI)
      *
      * @param string $apiKey
-     * @param CacheInterface $cache
      * @return Client
      */
-    static public function create($apiKey, $cache = null)
+    static public function create(string $apiKey)
     {
         $requestFactory = new RequestFactory();
         $rowFactory = new RowFactory();
         $resultFactory = new ResultFactory($rowFactory);
         $responseParser = new ResponseParser();
         $urlBuilder = new UrlBuilder();
-        $guzzle = new Guzzle();
-
-        $client = new Client($apiKey, $requestFactory, $resultFactory, $responseParser, $urlBuilder, $guzzle);
-
-        if (isset($cache)) {
-            $client->setCache($cache);
-        }
-
-        return $client;
+        $guzzle = new GuzzleClient();
+        return new Client($apiKey, $requestFactory, $responseParser, $resultFactory, $urlBuilder, $guzzle);
     }
-
-} 
+}
